Fixes https://java.net/jira/browse/JAVASERVERFACES-3668, Move 2.3 CDI producers out of API packages
git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@14169 761efcf2-d34d-6b61-9145-99dcacc3edf1
