+++ jsf-ri/conf/share/javax.enterprise.inject.spi.Extension
@@ -1,4 +1,4 @@
 com.sun.faces.application.view.ViewScopeExtension
 com.sun.faces.flow.FlowCDIExtension
 com.sun.faces.flow.FlowDiscoveryCDIExtension
-com.sun.faces.FacesCDIExtension
+com.sun.faces.cdi.CdiExtension
