+++ jsf-ri/src/main/java/com/sun/faces/cdi/ExternalContextProducer.java
@@ -8,7 +8,7 @@
  * and Distribution License("CDDL") (collectively, the "License").  You
  * may not use this file except in compliance with the License.  You can
  * obtain a copy of the License at
- * https://glassfish.dev.java.net/public/CDDL+GPL_1_1.html
+ * https://glassfish.dev.java.net/public/CDDLGPL_1_1.html
  * or packager/legal/LICENSE.txt.  See the License for the specific
  * language governing permissions and limitations under the License.
  *
@@ -37,7 +37,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package javax.faces.context;
+package com.sun.faces.cdi;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
@@ -52,17 +52,19 @@ import javax.enterprise.inject.Default;
 import javax.enterprise.inject.spi.Bean;
 import javax.enterprise.inject.spi.InjectionPoint;
 import javax.enterprise.util.AnnotationLiteral;
+import javax.faces.context.ExternalContext;
+import javax.faces.context.FacesContext;
 
 /**
  * <p class="changed_added_2_3">
- * The FacesContextProducer is the CDI producer that allows injection of the
- * FacesContext using @Inject and allows EL resolving of #{facesContext}
+ * The ExternalContextProducer is the CDI producer that allows injection of the
+ * ExternalContext using @Inject and allows EL resolving of #{externalContext}
  * </p>
  *
  * @since 2.3
  * @see FacesContext
  */
-public class FacesContextProducer extends CdiProducer implements Bean<FacesContext> {
+public class ExternalContextProducer extends CdiProducer implements Bean<ExternalContext> {
 
     /**
      * Inner class defining an annotation literal for @Default.
@@ -80,24 +82,19 @@ public class FacesContextProducer extends CdiProducer implements Bean<FacesConte
      * @return the Faces context.
      */
     @Override
-    public FacesContext create(CreationalContext<FacesContext> creationalContext) {
+    public ExternalContext create(CreationalContext<ExternalContext> creationalContext) {
         checkActive();
-        return FacesContext.getCurrentInstance();
+        return FacesContext.getCurrentInstance().getExternalContext();
     }
 
     /**
      * Destroy the instance.
      *
-     * <p>
-     * Since the FacesContext is a JSF artifact that the JSF runtime really is
-     * managing the destroy method here does not need to do anything.
-     * </p>
-     *
      * @param instance the instance.
      * @param creationalContext the creational context.
      */
     @Override
-    public void destroy(FacesContext instance, CreationalContext<FacesContext> creationalContext) {
+    public void destroy(ExternalContext instance, CreationalContext<ExternalContext> creationalContext) {
     }
 
     /**
@@ -127,7 +124,7 @@ public class FacesContextProducer extends CdiProducer implements Bean<FacesConte
      */
     @Override
     public String getName() {
-        return "facesContext";
+        return "externalContext";
     }
 
     /**
@@ -167,7 +164,7 @@ public class FacesContextProducer extends CdiProducer implements Bean<FacesConte
      */
     @Override
     public Set<Type> getTypes() {
-        return new HashSet<>(asList(FacesContext.class));
+        return new HashSet<>(asList(ExternalContext.class));
     }
 
     /**
